packages/chatterbox-tts/requirements.txt
chatterbox-tts
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
torch
torchaudio
numpy
